replicaset: fix net.box connection leak on detach
The net.box connection could not be garbage collected until
vconnect.future async call is garbage collected, since the
connection is referenced as a part of its async request.
The leak is reproducable only, when name_as_key
identification is used, since in other cases vconnect is not set.

Let's remove the future object from the net.box connection on
disconnect, so that it can be garbage collected after detach.

Part of tarantool/vshard#517

NO_DOC=bugfix