test: move acquiring the auto master cfg to the function
In the replicaset_3_test we have a lot of code duplication for
getting the auto master configuration. And the subsequent commit
also needs that.

Let's refactor the test a little bit, so that now there's a
function, which creates auto master global configuration.

Needed for tarantool/vshard#537

NO_DOC=test