replicaset: fix master not been found, when old one hangs
If instance hangs, then connection to it remains in 'active' state,
but we cannot make calls to it anymore, even pings doesn't work.
Replicaset's locate_master function assumped, that if master
doesn't respond, that means, that we cannot check for others,
since we don't know, whether the old master is active or not.

This is wrong assumtion and it's better to try searching for a new
master, since otherwise vshard won't be able to switch master at
all. Moreover, the protection of several masters in vshard is weak
(see, that we skip failed replicas during master search), so
there's nothing bad in trying to search for a new one.

It's important, that if old master does not respond and we're
unable to find a new one in replicaset, old one should not be
reset.

Closes tarantool/vshard#537

NO_DOC=bugfix