storage: fix route_map not been cleared
All deleted buckets are saved in the `route_map` to redirect routers
that arrive at the wrong replicaset. However, before this patch, the
map was cleared once every 500 years. The issue was that if the map
is empty, the clear deadline was set to infinity.

Let's set the timeout for cache to be `BUCKET_SENT_GARBAGE_DELAY` and
not `TIMEOUT_INFINITY`, as it was before the patch.

Note, that deadline in the patch have been replaced with timeout in
order to allow testing. Otherwise, as soon as deadline is set with big
timeout, it cannot be reset.

Additionally, the `BUCKET_SENT_GARBAGE_DELAY` constant has been
increased from 0.5 seconds to 60. It doesn't make sense to save
the cache just for half of the second: if GC works actively, the
cache will be cleared almost immediately.

Closes tarantool/vshard#546

NO_DOC=bugfix