storage: minimize the access of space during ref
Refs are used to protect buckets from rebalancing and GC. They're
taken on every user request and released after it. This is the most
hot place in vshard.

Ref has an optimization that if there is already > 0 refs of
the requested type, it works as an increment. But if the current
ref count is 0, it makes _bucket:get() to check bucket state.
That might be expensive - a tuple is returned to Lua as cdata,
which increases GC pressure. Also it might be relatively slow if
bucket count is millions.

We can completely get rid of accessing the space, when ref
exists. For that this commit makes the ref locks to be automatically
set, when they're needed.

Closes tarantool/vshard#285
Related tarantool/vshard#229

NO_DOC=refactoring