test: fix flaky router/router.test.lua
The part of the router/router test, where it's checked, that write
request for `SENDING` bucket fails, has become very flaky after
the implementation of the tarantool/vshard#285. Before the
patch there was TRANSFER_IN_PROGRESS error, now - BUCKET_IS_LOCKED.

However, this BUCKET_IS_LOCKED is retryable, TRANSFER_IN_PROGRESS
is not. This causes the flakiness of the test, since now we can get
either TimedOut or BUCKET_IS_LOCKED error. Let's change the test, so
that it becomes expected behavior.

It's also required to reset the bucket, since all of the subsequent
tests expect this bucket location to be unknown.

Follow-up tarantool/vshard#285

NO_DOC=test