test: fix flaky rebalancer/stress_add_remove_several_rs
At the end test drops the weight of one replicaset to 0 and
checks, that all other replicasets has exactly 100 buckets. However,
we didn't wait for rebalancing to end, which causes the flakiness:
it may happen, that not all buckets are ACTIVE, when we check them.

Let's wait for rebalancing to end and check the states of buckets
only after that.

NO_DOC=test