router: retry on TRANSFER_IS_IN_PROGRESS error
A retry is only illegal when we can't say if it was already applied or
not. TRANSFER_IS_IN_PROGRESS states that the request wasn't applied. So we
can  we can safely retry it.

Let's make the TRANSFER_IS_IN_PROGRESS consistent with BUCKET_IS_LOCKED
error, they both should be retryable.

Closes tarantool/vshard#548
Follow-up tarantool/vshard#285

NO_DOC=bugfix