replicaset: make locating master faster
When the old master hangs, we firstly make a call to it to check if it's
still alive. After that, we contact each replica in the replicaset and
wait for their responses. As a result, we end up calling the dead master
twice, which causes the master discovery process to lag by 3 *
MASTER_SEARCH_TIMEOUT seconds (2 times in `locate_master()` while
waiting for responses from dead master and 1 before waking up of the
master search fiber).

We can reduce the time required to discover the new master by skipping
the dead old master during the iteration over replicas. This would limit
the delay on the router to a maximum of 2 * MASTER_SEARCH_TIMEOUT per
master search iteration, if only one node is down.

Closes tarantool/vshard#549

NO_DOC=bugfix