router: introduce failover_sequential_fail_count option
This commit adds option for the bugfix, introduced in the
commit 9fc976daa04d6 ("router: calls affect temporary
prioritized replica"). It introduced marking replica as
unhealthy, when some number of requests failed in a row.

Follow-ups tarantool/vshard#505

@TarantoolBot document
Title: Document failover_sequential_fail_count router option

Router's failover constantly pings replicas to know, if they're
alive and properly working. If replica doesn't work correctly,
then its priority is temporary lowered and `callro` starts to
go to another replica. Moreover, callbro and other requests
skip this replica during balance, if it's unhealthy.
One of the reasons to mark replica as unhealthy is failed
requests in a row. This option allows to specify the number
of such failed requests to mark replica as unhealthy. Note,
that only network problems are considered as failed request.

The option is specified in the root of the config. Default value
is 3.