router: introduce failover_interval option
Follow-ups tarantool/vshard#505

@TarantoolBot document
Title: Document failover_interval router option

The interval in seconds between actions by the failover service.
For replicas it means, how often pings are sent, for replicasets -
how often the prioritized replica is checked for relevance.

The option is specified in the root of the config. Default value
is 1.