replicaset: fix replica marked as unhealthy on every replication flack
Before this patch, if upstream was in 'stopped' or 'disconnected'
state, replicas immediately was marked as unhealthy. This means,
if for upstream was down for 1 second, vshard could notice that
and switch the prioritized replica. We should rely only on the
`idle` and `lag` and not on the status of the upstream.

Part of tarantool/vshard#512

NO_DOC=bugfix