router: fix refs not being canceled on error in partial map_callrw
It all started with the major flakiness of the `router-luatest/map_callrw`
test, which was caused by the fact that `test_map_part_ref_timeout`
sometimes managed to execute the map step for `replica_1_a`. This happened
because the test relied on the assumption that the router had the
correct cache at startup — which was not the case. The previous test had
moved the bucket but did not ensure that the cache was updated
accordingly.

When the cache incorrectly pointed to rs2 for the bid2 bucket, the ref
part of `map_callrw` did not check whether a reference existed for rs2
twice, and thus failed to find the error and did map stage.

After fixing the flakiness, the test began to fail consistently due to
references not being deleted after a failed `map_callrw`. This was caused
by the fact that in replicasets_map_cancel_refs, future objects were
traversed — but those objects change with every iteration during the ref
stage of the partial `map_callrw`. The solution is to traverse all
replicasets, used in the `map_callrw`, instead.

Closes tarantool/vshard#508
Closes tarantool/vshard#554

NO_DOC=bugfix