vshard: allow stopping services on hot-reload
The hot-reload process was written as though the service will always
work in the file, where it was initially placed. However, this is not
the case, since in the subsequent commits several services will be moved
from storage and router to the replicaset.

The problem with the current implementation of the hot-reload is that,
if it happens from the version, where background fiber should work (new
one) to the one, where it shouldn't (old one), there's no way for
fiber to figure out, whether it should stop its work: even though
the `module_version` is bumped, the fiber function won't be updated,
since the `M is not updated on the old version for that specific
function and the function remains loaded in the `M`.

We need to unload all of the background functions and set them back on
every reload to solve that problem. Such behavior is implemented in this
commit. This commit allows for background services to exit in router,
storage, replicaset and util modules.

Needed for tarantool/vshard#505

NO_DOC=internal
NO_TEST=later