replicaset: introduce worker fiber
This patch adds worker, which will be used in the subsequent
commits to run replicaset services. The idea is to move failover,
master_search and conn_manager to the place, where they belong, in
the replicaset module.

Note, that hot-reload from the new version, with worker, to the one,
whithout it, requires either restart or manual stopping of the workers
before reload. Otherwise, several services will work, which will
litter the logs.

Needed for tarantool/vshard#505

NO_DOC=internal