router: move failover service to replicaset
Failover is moved to the replicaset module and it is divided into
several services: one for replicaset (increase/decrease priority of the
closest replica), other for replica (pings).

This commit also splits ERRINJ_FAILOVER_DELAY to several error
injections, since now it's required to stop services in replicas and
replicasets and this should not be done via single injection.

Also the implementation of the replicaset related part is simplified so
that it doesn't check, whether we have any work to be done, before the
patch we just checked that twice, not needed.

Needed for tarantool/vshard#505
Part of tarantool/vshard#263

NO_DOC=internal