router: fix failover assertion failure during upgrade process
The commit e1c806e1d3d2 ("router: make failover check health of the
storages") introduced checking for replication state of the replicas in
pings, but it also added the assertion, which checks, that replica
returns the health state. However, this is not the case during cluster
upgrade.

The documentation for cluster upgrade recommends the following:
   <...>
   4. Upgrade router nodes one by one.
   5. Upgrade storage replica sets one by one.

So, when the router is already upgraded and a storage is not, the
assertion failed, causing a bunch of LuaJIT errors in logs. Let's fix
that by applying the 'unknown' state to replica, if health is not
returned.

Closes tarantool/vshard#556

NO_DOC=bugfix