replicaset: speed up single master search
Currently the router spends too much time, waiting for calls to exit,
which increases latency of the cluster, since if node doesn't respond,
we wait for MASTER_SEARCH_TIMEOUT = 5 sec for him. And then we wait
for every async request to other nodes sequentially, so if master
is the second in the list, and the first one hangs, it'll take 5 more
seconds to find the master.

We can fasten the process up, by checking the futures for their
readiness in the loop and exit, as soon as the first master is found.
This is done in that commit.

This way we won't find, if there're several masters in the replicaset,
but the patch decreases the latency, which is much more important, in my
opinion. We sacrifice security in favor of reducing the latency during
master change. The task of ensuring the uniqueness of the master in the
replicaset should lie on external tools, as Raft consensus algorithm or
configuration tool.

Closes tarantool/vshard#540

NO_DOC=bugfix