test: fix flaky router-luatest/router_3_3_test
This commit fixes the flakiness of the router_3_3_test. There were a lot
of different fails, but in general they all are related to one problem:
the tests didn't wait for failover to end properly, which caused skip of
replicas in `callbro` request.

In order to wait for failover to end we need to fix the statuses, which
were incorrectly updated in the service itself.

Closes tarantool/vshard#534

NO_DOC=testfix