test: fix flaky failover/cluster_changes
Failover service was moved from router to replicaset in commit
33a2455f681f ("router: move failover service to replicaset"). The test
waits for "All replicas are ok" message in logs in order not to get
`SUBOPTIMAL_REPLICA` warning. However, now we have to wait for N such
messages, where N is the number of replicasets (3 in our case).

The other problem is that these messages may be written before we
actually started to wait and `grep_log` won't find them.

The simplest solution would be to wait for the exact number of warnings
we want to see.

Closes tarantool/vshard#561

NO_DOC=test