router: fix discovery service being nullified too late
If user tries to reconfigure vshard very often, it may happen, that
discovery service didn't manage to notice, that it's cancelled, before
the new one is started. In such case the new fiber may encounter
assertion failure and will be restarted several times.

The approach, which makes the service nullification, when the current
one is the one stored globally was chosen, since with blocking service
cancellation reconfiguration of vshard may take huge amount of time.

Part of #9

NO_DOC=bugfix