test: fix router-luatest/reload_test not reloading vshard
Before this commit the test loaded only the latest version, so the test
ran 2 times on the latest commit and did not test the actual hot-reload.
It tried to start the router on the latest version, changed
`package.loaded` and that's it, `vshard.router` even wasn't required
after that.

Since we don't support hot-reload from new version to the old one (see
0.1.34 release notes), let's actually create the router on old version
from the first start and then reload it to the latest. For that let's
use `LUA_PATH` environment variable. The alternative of using
`package.path` was rejected, since it requires modifying
`instances/router.lua` file.

However, changing `LUA_PATH` so that it points to `vshard_copy`
directory is not enough. The test runs from the repository directory and
`cwd` has higher priority, than `package.path` (which is set to
`LUA_PATH`). To bypass that, let's move the vshard copy to the
`override` directory, it has priority over `cwd`.

This commit also replaces all usages of `vtest.router_cfg` with local
function, since `ivtest.clear_test_cfg_options` may not present on old
version.

We also need to wait for service to be stopped before running the test
after reload during master_search service. The alternative solution is
using `router.master_search_f` to determine on which version function
is running. It's nullified right on reload. But let's better test, that
old services are actually stopped as soon as they're not needed.

Closes tarantool/vshard-ee#11

Co-authored-by: Roman Gromov <r.gromov.official@gmail.com>

NO_DOC=testfix