replicaset: rename health_status to replication_status
Before this commit information about replication state of replica was
saved into `replica.health_status`. Let's create a separate variable for
it, since in the next commit `health_status` will be used to save the
health of the whole replica (including dead connection, failed requests
and backoff).

The replication status has to be saved into the separate variable, since
the decision, whether it should affect the health of a replica, can be
made only by checking the replication status of the whole replicaset.
Consider the following case: 3 nodes, 1 of which is master and it's
down. Other 2 replicas doesn't have healthy replication to the master
but they anyway should be considered healthy, since we can and should
process RO requests.

Needed for tarantool/vshard#56

NO_DOC=refactoring