replicaset: speed up the read-only calls
Commit eb810a881a4f6ee39972 ("replicaset: skip unhealthy replicas in
all requests") caused major degradation of performance in read-only
calls, because it added the bunch of checks for every call when
searching for the healthy replica.

This commit makes updating health asynchronous by moving the status
check to failover service. Since we need to have access for the health
of all replicas, it's located in the replicaset's failover.

Even though the update of health status is asynchronous, in some cases
we can update it right away: backoff and sequential fails. And not only
we can, it has to be updated as soon as possible, since the speed of a
replica being moved to RED status affects the latency of the user
directly. But we cannot do that for connection being down fot N seconds
or for replication checks, so they remain asynchronous. That behavior is
already tested in router/router2.test and
router-luatest/router_3_3_test.

Closes #568

NO_DOC=bugfix