replicaset: fix read-only requests not falling back to master
The commit d017a9ed18767 ("replicaset: speed up the read-only calls")
has broken the fallback to master in read-only calls if master was in
backoff. Previoulsy, we checked the backoff and dropped it right during
calls, but after that commit only failover service started to do that,
which caused request failure even if backoff interval already passed,
because only on failover wakeup the instance could be restored.

This should not work like that and in case of fallback to master in
read-only calls we should explicitly check and drop backoff, if it's
possible, so that request can pass.

I also had an idea to restore the `health_status` of the replica, if
`replication_status` is `GREEN` and drop backoff right after every
successful call, but we should not do that, since a lot of services use
replicaset:call internally, when they don't want to take refs for
buckets and the success of such call doesn't mean, that vshard is
already configured and instance can be used for read-only calls. So, I
rejected that idea.

Closes #581

NO_DOC=bugfix