router: forbid using cdata as bucket_id
Using cdata (e.g. 1ULL) as bucket id was allowed in all requests of the
router. But using it caused the growth of the `route_map` table on every
request and full scan of the cluster in order to find, where the bucket
is. This happened due to the fact, that cdata is compared by its pointer,
not the value, and consequently, as there was no such pointer key in the
`route_map`, it was added and scanned every time.

It was decided to prohibit such usage, from now on we expect users to do
`tonumber(bucket_id)`, if it's needed. The motivation for such solution
is the fact, that all functions except calls already prohibit using
cdata as bucket_id, and calls should be consistent with it. Another
point is the fact, that the behavior of the `map_callrw` with cdata as
bucket ids is very untrivial.

The alternative solution of allowing usage of cdata in calls and manually
converting values to number in `bucket_set`, `bucket_reset` and
`bucket_resolved` was rejected due to the above mentioned reasons.

Unfortunately, the behavior of `map_callrw` with array of cdata is still
untrivial, as {10ULL, 20ULL} is interpreted as {1 = 10, 2 = 20}, and the
request is executed on the replicaset with 1st and 2nd bucket.

Closes #594

NO_DOC=bugfix