replicaset: do not spam with errors in logs on every failed call
Users are worried about the space on disk, which logs might use in case
of instance failure and we should relax the verbosity of logs for that.

`replica_call` in the replicaset module logged the error every time a
request fail, which produced a lot of logs, when instance is down. From
now on the same errors won't be logged, but we still need to know, when
instance becomes available, so after the first successful call new line
is logged now: "Calling 'func' on 'replica' succeeded".

Note, that comparing of errors messages cannot be done with _eq
metatable, since error is cdata, which is compared by pointers, so the
new method for error equality is introduced.

Part of #588

NO_DOC=bugfix