storage: deduplicate error messages in recovery service
Recovery service used to emit an error on every iteration while an
instance was unreachable, flooding logs with the same message. From
now on this won't be the case, same errors are not printed anymore.

This commit also almost complely get rids of `is_step_empty` variable
in the recovery service, but it's still needed unfortunately in order
to deduplicate the info about a bucket, if it wasn't recovered.

The new test file is introduced, since the tests are slow and I'd like
for them to run in parallel with others.

Part of #588

NO_DOC=bugfix