storage: relax rebalancer verbosity
Rebalancer service used to emit an error on every iteration while an
instance was unreachable, flooding logs with the same message. From
now on this won't be the case, same errors are not printed anymore.

`Some buckets are not active` info message is also deduplicated in this
patch. But the info about the number of buckets in the cluster - isn't,
because it usually says, that there're doubled buckets in the cluster.

Part of #588

NO_DOC=bugfix