log_ratelimit: introduce the module
This commit introduces the module, which can (and will) be used for
deduplication of the log messages. It does that according to the entries
{type = <string>, code = <int>} passed to it. Using only code is not
enough for deduplicating of the error objects, since different types of
errors may have the same codes.

The module itself doesn't log the message, since then all logs will have
the same line and file, a lot of information will be lost. So, the
module just says, whether an entry can be printed.

Note, that for `Suppressed ...` message to appear, `flush()` must be
called. For that there exists background service, which wakes up every
consts.LOG_RATELIMIT_INTERVAL and flushes all saves limiters.

Part of #578
Part of #583
Part of #588

NO_DOC=internal