replicaset: do not spam with errors in logs on every failed call
Users worry about the space on disk, which logs might use in case
of instance failure, and we should relax the verbosity of logs for that.

`replica_call` in the replicaset module logged the error every time a
request fail, which produced a lot of logs, when instance is down. From
now on the errors with same type and code won't be logged for 30 seconds.

Part of #588

NO_DOC=bugfix