storage: fix old master not being able to find a new master
This bug is reproduced only with auto_master feature enabled. It
manifests itself in inability for an old master to find a new one,
when switch happens.

M.this_replicaset.master and M.is_master becomes incosistent and old
master starts to respond, that it's not a master, but the only master it
knows is itself. This causes the spam of logs on the new master,
services there stop. Old master doesn't respond correctly to router's
ping, while being in this state.

In the absence of tarantool/vshard#605 bug this state ends in 40 seconds
(REPLICA_NOACTIVITY_TIMEOUT + MASTER_SEARCH_IDLE_INTERVAL * 2). However,
since connection manager cannot GC the stale connect and consequently
master search service cannot drop the master, it prolongs to infinity.

Speaking of implementation, we don't explicitly change the
`M.current_replicaset.master` variable in order to preserve
encapsulation of that variable, `replicaset:update_master` is used
instead.

Closes #604

NO_DOC=bugfix