replicaset: fix connection not GCed after rebind
The connection is rebound in order not to reconnect to instance on every
reconfiguration. However, if connection is rebound, it cannot be garbage
collected until the `activity_ts` variable is bumped (e.g. during some
call to replica). However, e.g. on replicas this variable never bumped,
so if replica was a master and stopped being it, `activity_ts` variable
won't be bumped at all and consequently, the connection won't be garbage
collected ever.

Closes #605

NO_DOC=bugfix