tla/storage: make recovery use bucket generation
This commit should resolve the #214 issue. Works as follows:

Recovery uses that generation in order to distinguish, which bucket is
more recent, if it cannot find a bucket on the sender node. So, firstly,
the node goes to the sender, if there's a bucket with any state and
greater generation, local one is `GARBAGE`, we don't care about the
status here. If bucket generation is equal to the local one, we use the
old logic, if the bucket is missing from remote node, then fullscan all
masters of the cluster. When all of the nodes replied, if there exists
higher generation, the local is `GARBAGE`, `ACTIVE` otherwise.

Part of #214

NO_DOC=tla
NO_TEST=tla