replicaset: log replica in connection triggers
Before this patch, only the host and port were written to logs on
connection or disconnection. However, it was very difficult to figure
out to which replica we connected/disconnected, which made debugging the
logs frustrating.

Let's instead log the full replica object, which includes the replica id
and its URI.

NO_DOC=refactoring